update: 인증제 점수 테이블 변경에 따른 리팩터링

+++ src/main/kotlin/com/team/incube/gsmc/v3/domain/score/service/impl/CreateJlptScoreServiceImpl.kt
@@ -49,7 +49,7 @@ class CreateJlptScoreServiceImpl(
                         existingScore.copy(
                             status = ScoreStatus.PENDING,
                             sourceId = fileId,
-                            scoreValue = grade,
+                            scoreValue = grade.toDouble(),
                         ),
                     )
                 } else {
@@ -61,7 +61,7 @@ class CreateJlptScoreServiceImpl(
                             status = ScoreStatus.PENDING,
                             sourceId = fileId,
                             activityName = null,
-                            scoreValue = grade,
+                            scoreValue = grade.toDouble(),
                         ),
                     )
                 }
