+++ src/main/kotlin/com/team/incube/gsmc/v3/domain/score/service/impl/CreateReadAThonScoreServiceImpl.kt
@@ -43,7 +43,7 @@ class CreateReadAThonScoreServiceImpl(
                         existingScore.copy(
                             status = ScoreStatus.PENDING,
                             sourceId = fileId,
-                            scoreValue = grade,
+                            scoreValue = grade.toDouble(),
                         ),
                     )
                 } else {
@@ -55,7 +55,7 @@ class CreateReadAThonScoreServiceImpl(
                             status = ScoreStatus.PENDING,
                             sourceId = fileId,
                             activityName = null,
-                            scoreValue = grade,
+                            scoreValue = grade.toDouble(),
                         ),
                     )
                 }
